refactored beam utils
