+++ zenml/datasources/csv_datasource.py
@@ -14,41 +14,17 @@
 """CSv Datasource definition"""
 
 import os
-from typing import Callable
-from typing import Text, Any, Dict
+from typing import Callable, Text, Dict
 
 import apache_beam as beam
-from tfx.components.example_gen.base_example_gen_executor import _WriteSplit
 from tfx_bsl.coders import csv_decoder
 
-from zenml.components.data_gen import utils
 from zenml.datasources.base_datasource import BaseDatasource
 from zenml.logger import get_logger
 from zenml.utils import path_utils
+from zenml.utils.beam_utils import WriteToTFRecord
 
 logger = get_logger(__name__)
-
-
-@beam.ptransform_fn
-@beam.typehints.with_input_types(beam.typehints.Dict[Text, Any])
-@beam.typehints.with_output_types(beam.pvalue.PDone)
-def WriteToTFRecord(datapoints: Dict[Text, Any],
-                    schema: Dict[Text, Any],
-                    output_split_path: Text) -> beam.pvalue.PDone:
-    """Infers schema and writes to TFRecord"""
-    # Obtain the schema
-    if schema:
-        schema_dict = {k: utils.SCHEMA_MAPPING[v] for k, v in schema.items()}
-    else:
-        schema = (datapoints
-                  | 'Schema' >> beam.CombineGlobally(utils.DtypeInferrer()))
-        schema_dict = beam.pvalue.AsSingleton(schema)
-
-    return (datapoints
-            | 'ToTFExample' >> beam.Map(utils.append_tf_example, schema_dict)
-            | 'WriteToTFRecord' >> _WriteSplit(
-                output_split_path=output_split_path
-            ))
 
 
 class CSVDatasource(BaseDatasource):
